Support port-forwarding without Kubernetes CLI
Fixes: #173

This change updates the `rad expose` command to work without:

- Local Kubernetes CLI
- Local Kubernetes credentials

We use the AKS Azure SDK to retrieve 'monitoring' credentials for the
environment. Then we can use the Kubernetes port-forward support as a
library.

I also reworked the command to make it clearer what you're specifying
when you work with it. You also have the ability to specify the remote
port on the container, this wasn't supported before, it had to be
something you exposed as a service. This is much more flexible and
probably easier to understand as well.