Example "first-application" walkthrough
This is an example of a new intro for Radius that's meant to mirror the
Dapr getting started. This is just a paper prototype for now because we
haven't built all of the features that will power this experience.

This section would go after installation of the CLI/Tools and before the
Quickstarts and Tutorial sections. We would use this example as the the
thing we show users first to understand Radius.

Right now this is designed to highlight:

- Streamlined "dev" setup that asks you for less linput
- A very simple way to deploy a container (simpler than today)
- Streaming logs and automatic port-forwarding