minor language changes in tutorial (#384)
* change a section header in Author app page

to add clarity about the fact that this section is about the Radius app resource. based on user feedback.

* keep platform name consistent across page

currently we say "radius env" and "radius k8s env" and a user asked if they're the same thing

* remove duplicative namespace callout 

based on user feedback:
> this callout is redundant and the content is self-explained by the code underneath

* typo fix on Limitations page

* move note about accessing app source code closer to app diagram

users were wondering where it was. I guess it got lost in the resource description text before.

* Update docs/content/getting-started/tutorial/1-tutorial-environment/index.md

Co-authored-by: Aaron Crawfis <Aaron.Crawfis@microsoft.com>

Co-authored-by: Aaron Crawfis <Aaron.Crawfis@microsoft.com>