Add API concepts and update architecture
This change adds a new API 'concepts' page and fleshes out the existing architecture concepts page. We currently have a bad|old|outdated|wrong version of something like this as part of the contributing docs in `project-radius/radius`. My plan is to suppliment the official docs with the correct info and link to it from the in-repo contributing docs.